Prevent update of date if it didn't changed
The "dp.change" event is triggered on init, while the date didn't changed.

This issue comes when you bound the date to a model property. When the first "dp.change" event is called on init, the property is updated on the model and the isDirty property of the model becomes 'true'... whereas the value didn't changed.

This fix prevents that issue by checking if date value really changed.