BREAKING CHANGE: textField options are not supported anymore
Ember 2 will deprecate the view helper (http://emberjs.com/deprecations/v1.x/#toc_view-and-controller-template-keywords). The first two betas already did this. Ember 2 beta 3 goes further and drops it altogether. This makes the ember-bootstrap-datetimepicker incompatible with the latest versions of Ember 2 beta.

To make it compatible we have dropped to `textField*` option

If you want to configure the textfield inside the picker do this:

````handlebars
{{#bs-datetimepicker date=model.date2}}
  {{input class="form-control" name="Date" placeholder="Date"}}
{{/bs-datetimepicker}}

```